refactor: 🚚 Rename "ranksSeasons.ts" -> "ranks.ts"
